% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOBmodelmain.R
\name{argF}
\alias{argF}
\title{Title}
\usage{
argF(
  cohorts,
  predictors,
  asset_data,
  soil_data,
  val_start,
  val_end,
  test_start,
  test_end,
  outages,
  maxN,
  minN,
  Nfailcutoff,
  rfe,
  updateData,
  forceUpdate,
  mainDir
)
}
\arguments{
\item{mainDir}{}
}
\value{

}
\description{
Title
}
