% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangle.R
\name{coalesce2}
\alias{coalesce2}
\title{Modified Coalesce
Coalesce two dataframes}
\usage{
coalesce2(...)
}
\arguments{
\item{...}{}
}
\value{
dataframe
}
\description{
Modified Coalesce
Coalesce two dataframes
}
\examples{
\dontrun {coalesce2(x, y)}
}
