% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangle.R
\name{data_wrangle}
\alias{data_wrangle}
\title{Data Wrangle}
\usage{
data_wrangle(imported_data)
}
\arguments{
\item{imported_data}{list}
}
\value{
list of 2 cleaned data frames:  asset data and work order data
}
\description{
This is a custom function which creates two cleaned data frames
1: asset data and 2: a combined work order and asset data
The cleaning steps ensures there are no duplicates and that the data is consistent
}
\examples{
\dontrun{
data_wrangle(imported_data)
}
}
