% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H20.R
\name{GBMstackedensemble}
\alias{GBMstackedensemble}
\title{Trains a stacked ensemble model using H20}
\usage{
GBMstackedensemble(classification, dftrain, dftest, upsample, SMOTE, path)
}
\arguments{
\item{classification}{logical}

\item{dftrain}{dataframe}

\item{dftest}{dataframe}

\item{upsample}{logical}

\item{SMOTE}{logical}

\item{path}{string}
}
\value{
list of objects including performance, confusion Matrices, finalModel, Hyperparameter grid, bestModelID and model saved path)
}
\description{
Trains a stacked ensemble model using H20
}
\examples{
\dontrun{
GBMstackedensemble<-function(classification, dftrain, dftest, upsample, SMOTE, path)
}
}
