% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get.soil.data}
\alias{get.soil.data}
\title{GetSoil Data}
\usage{
get.soil.data(lat1, lon1, soil_data)
}
\arguments{
\item{lat1}{numeric}

\item{lon1}{numeric}

\item{soil_data}{dataframe}
}
\value{

}
\description{
This function gets soil data i.e. clay, sand ect... by longitude and latitude
}
\examples{
\dontrun{
get.soil.data(lat=37.8, lon1=-144)
}
}
