% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getBOM}
\alias{getBOM}
\title{getBOM
This function processes climate data from BOM including temp, rainfall,}
\usage{
getBOM(update, ClimatePath)
}
\arguments{
\item{update}{}
}
\value{
date frame of climate data
}
\description{
getBOM
This function processes climate data from BOM including temp, rainfall,
}
\examples{
\dontrun {getBOM(update=FALSE, ClimatePath="PathtoClimateData")}
}
