% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_data}
\alias{import_data}
\title{Import}
\usage{
import_data(
  AssetPath,
  WorkOrderPath,
  HansenPath,
  newWorkOrdersPath,
  rainfall,
  minTemp,
  maxTemp,
  soilData,
  GISdata,
  ClimatePath,
  SOBSLID
)
}
\arguments{
\item{AssetPath}{Path to Asset Data csv}

\item{WorkOrderPath}{Path to Work Order Maximo Data csv}

\item{HansenPath}{Path to Hansen Work Order Data csv}

\item{newWorkOrdersPath}{Path to New Maximo Work order Data csv}

\item{rainfall}{Path to rainfall grid data RDS}

\item{minTemp}{Path to min temp grid data RDS}

\item{maxTemp}{Path to max temp data RDS}

\item{soilData}{Path to Asset Data tif files}

\item{GISdata}{Path to GIS Data csv}

\item{ClimatePath}{character}

\item{SOBSLID}{dataframe}
}
\value{
list of dataframes
}
\description{
This function is to import all requird data sources for
}
\examples{
\dontrun{
import_data(AssetPath="~Data/assetData.csv")
}
}
