% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H20.R
\name{modelPreProc_update}
\alias{modelPreProc_update}
\title{Title}
\usage{
modelPreProc_update(
  combined.df,
  predictors,
  classification,
  Nfailcutoff,
  outlierRemove,
  trained_rec
)
}
\arguments{
\item{recPath}{}
}
\value{

}
\description{
Title
}
