% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOBmodelmain.R
\name{ModPred}
\alias{ModPred}
\title{Model Predict}
\usage{
ModPred(Predictions, Labels, predDF, Thr, newPosClassLabel, newNegClassLabel)
}
\arguments{
\item{newNegClassLabel}{}
}
\value{

}
\description{
Model Predict
}
