% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPPSOB.R
\name{NHPP_fit}
\alias{NHPP_fit}
\title{Title}
\usage{
NHPP_fit(
  minNinSOB,
  asset_data,
  soil_data,
  work_Order_Data,
  SOBNr,
  TI1,
  TI2,
  FNRupp,
  FNRlow,
  plot,
  minpkh,
  rollingwin,
  inclsoilmoist
)
}
\arguments{
\item{inclsoilmoist}{}
}
\value{

}
\description{
Title
}
