% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPPSOB_update.R
\name{NHPP_loadData4_update}
\alias{NHPP_loadData4_update}
\title{Title}
\usage{
NHPP_loadData4_update(
  mainOnly,
  val_start,
  work_Order_Data,
  asset_data,
  val_end,
  outages,
  maxNinSOB,
  minNinSOB,
  test_start,
  test_end,
  subDir
)
}
\arguments{
\item{subDir}{}
}
\value{

}
\description{
Title
}
