% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPPCohort.R
\name{NHPP_loadDatacohort}
\alias{NHPP_loadDatacohort}
\title{NHPP_loadDatacohort}
\usage{
NHPP_loadDatacohort(
  work_order,
  val_start,
  val_end,
  outages,
  maxN,
  minN,
  test_start,
  test_end
)
}
\arguments{
\item{work_order}{datafrane}

\item{val_start}{date}

\item{val_end}{date}

\item{outages}{logical}

\item{maxN}{integer}

\item{minN}{integer}

\item{test_start}{date}

\item{test_end}{date}
}
\value{
datafrane for NHPP Cohort Fitting
}
\description{
Analagous function to SOB, applied to pipe Cohorts
}
\examples{
\dontrun{
NHPP_loadDatacohort <- function(work_order, val_start, val_end, outages, maxN, minN, test_start, test_end)
}
}
