% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeAnalysis.R
\name{pipe_analysis}
\alias{pipe_analysis}
\title{Pipe AnalysisThis function groups and runs metrics on the pipes and work orders by cohort}
\usage{
pipe_analysis(outages, asset_workorder_data, update)
}
\arguments{
\item{update}{}
}
\value{

}
\description{
Pipe AnalysisThis function groups and runs metrics on the pipes and work orders by cohort
}
