% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipemodelMain.R
\name{PipeIntensity}
\alias{PipeIntensity}
\title{Title
Note this determines the value of alpha or intensity of failure, lambda(t) for pipes.  It assumes that alpha is constant with time and is updated based on new evidence, i.e  N and T}
\usage{
PipeIntensity(
  DZ,
  SOB,
  asset_data,
  work_Order_Data,
  cohortFailure,
  cohortNr,
  test_start,
  test_end
)
}
\arguments{
\item{DZ}{}

\item{test_end}{}
}
\value{

}
\description{
Title
Note this determines the value of alpha or intensity of failure, lambda(t) for pipes.  It assumes that alpha is constant with time and is updated based on new evidence, i.e  N and T
}
