% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipemodelMain.R
\name{pipeModel}
\alias{pipeModel}
\title{PipeModel
Individual pipe intensity modelling}
\usage{
pipeModel(
  workorder_data,
  asset_data,
  cohort_data,
  SOBData,
  WDZAssetPath,
  Zones,
  SOBResult,
  NHPPResult,
  ResultPath,
  TestStart,
  TestEnd
)
}
\arguments{
\item{TestEnd}{}
}
\value{

}
\description{
PipeModel
Individual pipe intensity modelling
}
