% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeAnalysis.R
\name{preCohort}
\alias{preCohort}
\title{CohortAnalysis
this function updates the cohorts and runs analysis and metrics on the data}
\usage{
preCohort(update, outages)
}
\arguments{
\item{outages}{}
}
\value{

}
\description{
CohortAnalysis
this function updates the cohorts and runs analysis and metrics on the data
}
