% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipemodelMain.R
\name{rankWDZ}
\alias{rankWDZ}
\title{rankWDZ}
\usage{
rankWDZ(WDZPath, SOBResult, N, newPosClassLabel)
}
\arguments{
\item{WDZPath}{character}

\item{SOBResult}{dataframe}

\item{N}{integer}
}
\value{
a list of the top N Zones
}
\description{
ranks WDZ in order of decreasing proportion of SOBs
}
\examples{
\dontrun{
rankWDZ <- function(WDZPath, SOBResult, N=10)
}
}
