% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOBPredictR-package.R
\docType{package}
\name{SOBPredictR-package}
\alias{SOBPredictR}
\alias{SOBPredictR-package}
\title{SOBPredictR: Analysis and Predictive Modelling of Water Main Failures}
\description{
"Imports and analyses asset and work order information in order to predict whether a Shut Off Block will experience a failure"
}
\author{
\strong{Maintainer}: Brendan Gladman \email{b.gladman@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
