% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipemodelMain.R
\name{ZoneResults}
\alias{ZoneResults}
\title{Zone Results}
\usage{
ZoneResults(PipeResults, PrCutOff, workorder_data, maxLength)
}
\arguments{
\item{PipeResults}{dataframe}

\item{PrCutOff}{numeric}

\item{workorder_data}{dataframe}

\item{maxLength}{numeric}
}
\value{
list of dataframes
}
\description{
Pipe Modelling for Zones
}
\examples{
\dontrun{
ZoneResults<-function(PipeResults, PrCutOff, workorder_data, maxLength)
}
}
