% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangle.R
\name{coalesce2}
\alias{coalesce2}
\title{Modified Coalesce}
\usage{
coalesce2(...)
}
\arguments{
\item{...}{vector}
}
\value{
vector
}
\description{
Coalesce two dataframes
}
\examples{
\dontrun{
coalesce2(x, y)
}
}
