% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H20.R
\name{modelPreProc}
\alias{modelPreProc}
\title{Model PreProcess}
\usage{
modelPreProc(
  combined.df,
  predictors,
  classification,
  Nfailcutoff,
  outlierRemove
)
}
\arguments{
\item{combined.df}{dataframe}

\item{predictors}{character}

\item{classification}{logical}

\item{Nfailcutoff}{integer}

\item{outlierRemove}{logical}
}
\value{

}
\description{
PreProcess data for training a new SOB model
}
\examples{
\dontrun{
modelPreProc<- function(combined.df, predictors, classification, Nfailcutoff,  outlierRemove)
}
}
