% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPPCohort.R
\name{NHPP_fit_cohort}
\alias{NHPP_fit_cohort}
\title{NHPP_fit_cohort}
\usage{
NHPP_fit_cohort(
  asset_data,
  work_Order_Data,
  cohortNr,
  TI1,
  TI2,
  FNRupp,
  FNRlow,
  plot,
  minpkh,
  rollingwin,
  inclsoilmoist
)
}
\arguments{
\item{asset_data}{dataframe}

\item{work_Order_Data}{dataframe}

\item{cohortNr}{integer}

\item{TI1}{integer}

\item{TI2}{integer}

\item{FNRupp}{numeric}

\item{FNRlow}{numeric}

\item{plot}{logical}

\item{minpkh}{numeric}

\item{rollingwin}{integer}

\item{inclsoilmoist}{logical}
}
\value{
dataframe of NHPP fitted parameters for all pipe Cohorts.  This is required for the NHPP
}
\description{
Analagous function to SOB NHPP fit.  Note this function is not called directly but via the modelTrain function
}
\examples{
\dontrun{
NHPP_fit_cohort <- function(asset_data, work_Order_Data, cohortNr, TI1, TI2, FNRupp, FNRlow, plot, minpkh, rollingwin, inclsoilmoist)
}
}
