% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPPCohort.R
\name{NHPP_fit_cohort_update}
\alias{NHPP_fit_cohort_update}
\title{Title}
\usage{
NHPP_fit_cohort_update(
  asset_data,
  work_Order_Data,
  cohortNr,
  TI1,
  TI2,
  FNRupp,
  FNRlow,
  plot,
  minpkh,
  rollingwin,
  inclsoilmoist
)
}
\arguments{
\item{inclsoilmoist}{}
}
\value{

}
\description{
Title
}
