% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPPCohort.R
\name{NHPP_loadDatacohort}
\alias{NHPP_loadDatacohort}
\title{Title}
\usage{
NHPP_loadDatacohort(
  work_order,
  val_start,
  val_end,
  outages,
  maxN,
  minN,
  test_start,
  test_end
)
}
\arguments{
\item{test_end}{}
}
\value{

}
\description{
Title
}
