% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeAnalysis.R
\name{pipe_cohort_analysis}
\alias{pipe_cohort_analysis}
\title{Pipe AnalysisThis function groups and runs metrics on the pipes and work orders by cohort}
\usage{
pipe_cohort_analysis(outages, workorder_data, asset_data)
}
\arguments{
\item{outages}{}

\item{workorder_data}{}
}
\value{

}
\description{
Pipe AnalysisThis function groups and runs metrics on the pipes and work orders by cohort
}
