% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipemodelMain.R
\name{PipeIntensity}
\alias{PipeIntensity}
\title{PipeIntensity}
\usage{
PipeIntensity(
  DZ,
  SOB,
  asset_data,
  work_Order_Data,
  CohortID,
  cohortFailure,
  test_start,
  test_end
)
}
\arguments{
\item{DZ}{character}

\item{SOB}{integer}

\item{asset_data}{dataframe}

\item{work_Order_Data}{dataframe}

\item{CohortID}{integer}

\item{cohortFailure}{dataframe}

\item{test_start}{date}

\item{test_end}{date}
}
\value{

}
\description{
This determines the value of alpha or intensity of failure, lambda(t) for pipes.  It assumes that alpha is constant with time and is updated based on new evidence, i.e  N and T
}
\examples{
\dontrun{
PipeIntensity <- function(DZ, SOB, asset_data, work_Order_Data, CohortID, cohortFailure,
 test_start,test_end)
}
}
