% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOBmodelmain.R
\name{SOBmodelPredict}
\alias{SOBmodelPredict}
\title{Title}
\usage{
SOBmodelPredict(
  workorder_data,
  asset_data,
  SOB_data,
  soil_data,
  val_start,
  val_end,
  test_start,
  test_end,
  outages = FALSE,
  predictors,
  ModelObj
)
}
\arguments{
\item{savePaths}{}
}
\value{

}
\description{
Title
}
