% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPPSOB.R
\name{SOBNHPP_load}
\alias{SOBNHPP_load}
\title{Title}
\usage{
SOBNHPP_load(
  mainOnly,
  val_start,
  val_end,
  outages,
  maxNinSOB,
  work_Order_Data,
  asset_data,
  SOB_data,
  minNinSOB,
  test_start,
  test_end,
  subDir,
  cohortResults
)
}
\arguments{
\item{cohortResults}{}
}
\value{

}
\description{
Title
}
